Finishes deployment to mobile and heroku
Unfortunately, for click events to the center marker to work on mobile, the google map needed to be in a div that was not constructed by a directive. Moves the instantiation of the map and its listeners to the MapCtrl controller and gets rid of the map directive as well as all of its references.

Also, remakes the package.json to run bower install as a post install script on heroku

Fixes loading of map when coming from homepage. This causes the controller to be much much bigger than it ought to be, so condensing of functionality into services will be the next step

Adds logo to README.md